Handle recursive creation of pointers
